Move i18n eslint rules to package
