+++ packages/grafana-eslint-rules/index.cjs
@@ -1,8 +1,6 @@
 const noAriaLabelSelectors = require('./rules/no-aria-label-e2e-selectors.cjs');
 const noBorderRadiusLiteral = require('./rules/no-border-radius-literal.cjs');
 const noUnreducedMotion = require('./rules/no-unreduced-motion.cjs');
-const noUntranslatedStrings = require('./rules/no-untranslated-strings.cjs');
-const noTranslationTopLevel = require('./rules/no-translation-top-level.cjs');
 const themeTokenUsage = require('./rules/theme-token-usage.cjs');
 const noRestrictedImgSrcs = require('./rules/no-restricted-img-srcs.cjs');
 
@@ -12,8 +10,6 @@ module.exports = {
     'no-aria-label-selectors': noAriaLabelSelectors,
     'no-border-radius-literal': noBorderRadiusLiteral,
     'theme-token-usage': themeTokenUsage,
-    'no-untranslated-strings': noUntranslatedStrings,
-    'no-translation-top-level': noTranslationTopLevel,
     'no-restricted-img-srcs': noRestrictedImgSrcs,
   },
 };
