+++ packages/grafana-i18n/src/rules/no-translation-top-level/no-translation-top-level.cjs
